# backend/tracker/docs/swagger.yaml
openapi: 3.0.0
info:
  title: Nutrition Tracker API
  version: 1.0.0
  description: |
    The Tracker API allows users to manage their products, consumption logs, nutritional statistics, shopping lists, and personal settings. 
    This API provides various endpoints for retrieving product details, adding or removing items from shopping lists, consuming products and tracking daily nutrition, 
    updating personal settings, and more.

    Key functionalities include:
    - Managing products: Get, create, and search products by ID or name.
    - Consuming products: Record product consumption and update nutrition statistics.
    - Managing shopping lists: Create, retrieve, and delete shopping lists, as well as add or remove items from them.
    - Tracking daily nutrition: Retrieve detailed nutrition statistics based on consumed products.
    - Personal settings: View and update user-specific nutritional goals and preferences.
    - Calendar integration: View, create, and update calendar events related to food consumption.
...
servers:
  - url: http://localhost:9401/tracker
    description: "standard url"
  - url: http://tracker-api:9401/tracker
    description: "url when fetching data from this api inside a docker network"
paths:
  /products:
    get:
      summary: List all products
      description: Retrieves a list of all available products.
      tags:
        - products
      operationId: getProducts
      responses:
        '200':
          description: A list of products.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductDTO'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
                  example: "Failed to retrieve product."
  /products/{id}:
    get:
      summary: Get a product by ID
      description: Retrieves details of a specific product using its ID.
      tags:
        - products
      operationId: getProductByID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Product details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDTO'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
                  example: "Product not found."
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
                  example: "Failed to retrieve product."
  /products/name/{name}:
    get:
      summary: Search products by name
      description: Retrieves a list of products matching a given name.
      tags:
        - products
      operationId: getProductByName
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: A list of matching products.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductDTO'
        '404':
          description: No products found with the given name
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
                  example: "Product not found."
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
                  example: "Failed to retrieve product."
  /consume:
    post:
      summary: Consume a product and update nutrition statistics.
      description: This endpoint allows a user to consume a product and update the nutrition statistics for that product.
      tags:
        - consume
      operationId: consumeProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExpectedProductDTO'
      responses:
        201:
          description: Product consumed and nutrition statistics updated successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  consumedProduct:
                    $ref: '#/components/schemas/ConsumedProduct'
        400:
          description: Bad request. Invalid input data.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Invalid data format."
        409:
          description: Product ID already exists.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Product ID already exists."
        500:
          description: Internal server error. Something went wrong while processing the request.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Failed to create/retrieve daily record."
  /diary/date/{date}:
    get:
      summary: Retrieve the diary of consumed products for a specific date.
      description: Fetches the list of consumed products for a given date.
      tags:
        - diary
      operationId: getDiaryByDate
      parameters:
        - name: date
          in: path
          required: true
          description: The date for which to retrieve the diary of consumed products (format YYYY-MM-DD).
          schema:
            type: string
            format: date
      responses:
        200:
          description: Successfully retrieved the diary for the specified date.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DailyProductsConsumed'
        400:
          description: Invalid date format.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Invalid date format. Use YYYY-MM-DD."
        404:
          description: No records found for the given date.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "No records found for this date."
        500:
          description: Internal server error when fetching diary data.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Failed to fetch diary entry."
  /diary/date/{date}/{id}:
    get:
      summary: Retrieve a specific consumed product by its ID.
      description: Fetches details of a consumed product based on its ID for a given date.
      tags:
        - diary
      operationId: getConsumedProductByID
      parameters:
        - name: date
          in: path
          required: true
          description: The date for which to retrieve the product (format YYYY-MM-DD).
          schema:
            type: string
            format: date
        - name: id
          in: path
          required: true
          description: The ID of the consumed product.
          schema:
            type: integer
      responses:
        200:
          description: Successfully retrieved the consumed product.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsumedProduct'
        400:
          description: Invalid ID format (must be an integer).
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Invalid ID format. Must be an integer."
        404:
          description: Product not found for the given ID.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Product not found."
        500:
          description: Internal server error while fetching the consumed product.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Failed to fetch product."
    delete:
      summary: Delete a specific consumed product.
      description: Deletes a consumed product by its ID and updates nutrition statistics.
      tags:
        - diary
      operationId: deleteConsumedProduct
      parameters:
        - name: date
          in: path
          required: true
          description: The date for which the product is being deleted (format YYYY-MM-DD).
          schema:
            type: string
            format: date
        - name: id
          in: path
          required: true
          description: The ID of the consumed product to delete.
          schema:
            type: integer
      responses:
        200:
          description: Successfully deleted the product and updated nutrition statistics.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Product deleted successfully."
        400:
          description: Invalid ID format (must be an integer).
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Invalid ID format. Must be an integer."
        404:
          description: Product not found for the given ID.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Product not found."
        500:
          description: Internal server error when deleting the product or updating nutrition statistics.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Failed to delete product."
  /settings:
    get:
      summary: Retrieve the current settings.
      description: Fetches the current user settings for nutritional goals and preferences.
      tags:
        - settings
      operationId: getSettings
      responses:
        200:
          description: Successfully retrieved the current settings.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Settings'
        404:
          description: Settings not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Settings not found."
        500:
          description: Internal server error while fetching the settings.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Failed to fetch settings."
    patch:
      summary: Update the user settings.
      description: Updates the user's nutritional goals and preferences based on the provided data.
      tags:
        - settings
      operationId: updateSettings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SettingsDTO'
      responses:
        200:
          description: Successfully updated the user settings.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        400:
          description: Invalid request body (e.g., missing required fields or incorrect data format).
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Invalid data format."
        500:
          description: Internal server error while updating the settings.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Failed to update settings."
  /nutrition/{date}:
    get:
      summary: Retrieve the nutrition statistics for a specific date.
      description: Fetches the total nutrition statistics for a given date based on consumed products.
      tags:
        - nutrition
      operationId: getNutritionStatisticsByDate
      parameters:
        - name: date
          in: path
          required: true
          description: The date for which to retrieve nutrition statistics (format YYYY-MM-DD).
          schema:
            type: string
            format: date
      responses:
        200:
          description: Successfully retrieved the nutrition statistics for the specified date.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NutritionStatistics'
        400:
          description: Invalid date format. Date must be in YYYY-MM-DD format.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Invalid date format. Use YYYY-MM-DD."
        500:
          description: Internal server error while fetching nutrition statistics.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Failed to fetch nutrition statistics."
  /calendar/{date}:
    get:
      summary: Retrieve calendar entries for a specific date.
      description: Fetches calendar entries for a given date.
      tags:
        - calendar
      operationId: getCalendarByDate
      parameters:
        - name: date
          in: path
          required: true
          description: The date for which to retrieve calendar entries (format YYYY-MM-DD).
          schema:
            type: string
            format: date
      responses:
        200:
          description: Successfully retrieved calendar entries.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CalendarDTO'
        400:
          description: Invalid date format.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Invalid date format. Use YYYY-MM-DD."
        500:
          description: Internal server error while fetching calendar entries.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Failed to fetch calendar entries."
  /calendar:
    post:
      summary: Add a new calendar entry.
      description: Creates a new calendar entry with date and time details.
      tags:
        - calendar
      operationId: addCalendarEntry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CalendarDTO'
      responses:
        201:
          description: Successfully added calendar entry.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Calendar entry added successfully."
        400:
          description: Invalid input data.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Invalid request payload."
        500:
          description: Internal server error while adding calendar entry.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Failed to add calendar entry."
  /calendar/{id}:
    patch:
      summary: Update an existing calendar entry.
      description: Modifies details of a calendar entry by ID.
      tags:
        - calendar
      operationId: updateCalendarEntry
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the calendar entry to update.
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CalendarDTO'
      responses:
        200:
          description: Successfully updated the calendar entry.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Calendar entry updated successfully."
        400:
          description: Invalid input data.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Invalid request payload."
        500:
          description: Internal server error while updating calendar entry.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Failed to update calendar entry."

    delete:
      summary: Delete a calendar entry.
      description: Removes a calendar entry by ID.
      tags:
        - calendar
      operationId: deleteCalendarEntry
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the calendar entry to delete.
          schema:
            type: integer
      responses:
        200:
          description: Successfully deleted the calendar entry.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Calendar entry deleted successfully."
        404:
          description: Calendar entry not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Calendar entry not found."
        500:
          description: Internal server error while deleting calendar entry.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Failed to delete calendar entry."
...
components:
  schemas:
    ProductDTO:
      type: object
      properties:
        ID:
          type: integer
          format: uint
        Name:
          type: string
        Brand:
          type: string
        CaloriesPer100Grams:
          type: integer
        ProteinsInGrams:
          type: number
          format: float
        FatsInGrams:
          type: number
          format: float
        CarbsInGrams:
          type: number
          format: float
    ExpectedProductDTO:
      type: object
      properties:
        Product:
          $ref: '#/components/schemas/ProductDTO'
        Date:
          type: string
        Weight:
          type: integer
        Category:
          type: string
    DailyProductsConsumed:
      type: object
      properties:
        ID:
          type: integer
        Date:
          type: string
        Products:
          type: array
          items:
            $ref: '#/components/schemas/ConsumedProduct'
    ConsumedProduct:
      type: object
      properties:
        ID:
          type: integer
          format: uint
        DailyProductsConsumedID:
          type: integer
        ProductID:
          type: integer
        ProductName:
          type: string
        Brand:
          type: string
        Category:
          type: string
          enum: [Frühstück, Mittagessen, Abendessen, Snack]
        WeightInGrams:
          type: integer
        Calories:
          type: integer
        ProteinsInGrams:
          type: number
          format: float
        FatsInGrams:
          type: number
          format: float
        CarbsInGrams:
          type: number
          format: float
    NutritionStatistics:
      type: object
      properties:
        ID:
          type: integer
          format: uint
        Date:
          type: string
        ConsumedCalories:
          type: integer
        ConsumedProteins:
          type: number
          format: float
        ConsumedFats:
          type: number
          format: float
        ConsumedCarbs:
          type: number
          format: float
    SettingsDTO:
      type: object
      properties:
        PlannedCalories:
          type: integer
        Gender:
          type: string
        WeightInKg:
          type: number
          format: float
    Settings:
      type: object
      properties:
        PlannedCalories:
          type: integer
        ProteinsInGrams:
          type: integer
        FatsInGrams:
          type: integer
        CarbsInGrams:
          type: integer
        Gender:
          type: string
        WeightInKg:
          type: number
          format: float
    CalendarDTO:
      type: object
      properties:
        ID:
          type: integer
        Date:
          type: string
        BeginTime:
          type: string
        EndTime:
          type: string
        Name:
          type: string
        Description:
          type: string
...